Bug fixing success dash+MovingPlatforms
I have removed old moving platform and platformer moving scripts as well as reintroducing the newer and hopefully fully fixed versions. dashing has returned, only being able to be used once reset when in contact with the ground (like jumping) and a bug has been removed where when getting off of a moving platform you still move with it. Due note; my fix for this isn't as good as I would like it to be, you will still move with the platform until you come into contact with other ground.